almost there..... nearly finsihed this one, just have to beat this one last test.

+++ largest-series-product/src/largest_series_product.h
@@ -1,6 +1,8 @@
 #ifndef LARGEST_SERIES_PRODUCT_H 
 #define LARGEST_SERIES_PRODUCT_H
 
-unsigned int largest_series_product (const char* numbers, int count);
+#include <stdint.h>
+
+int64_t largest_series_product (const char* numbers, unsigned int count);
 
 #endif
